Change configs, fix build errors and warnings on linux

+++ source/GameState.cpp
@@ -142,7 +142,7 @@ void GameState::StepLogEntry(char*& world_buffer) const
 	world_buffer += sprintf(world_buffer, "m %i %i\n", pacman.position.x, pacman.position.y);
 	for (size_t ghost_index = 0; ghost_index < ghosts.size(); ++ghost_index)
 	{
-		world_buffer += sprintf(world_buffer, "%i %i %i\n", ghost_index + 1, ghosts[ghost_index].position.x, ghosts[ghost_index].position.y);
+		world_buffer += sprintf(world_buffer, "%i %i %i\n", static_cast<int>(ghost_index + 1), ghosts[ghost_index].position.x, ghosts[ghost_index].position.y);
 	}
 }
